NAV-24802: Splitter opp LeggTilEllerFjernBrevmottakere i to komponenter, en for behandling og en for fagsak
